Fully Stable Build

+++ BeneathTheSurface/src/Level.java
@@ -15,6 +15,6 @@ public class Level
 		objectsInWorld[1] = new Object(20, Runner.sHeight - 55, Runner.sWidth, 20, true, false, "ground");
 		objectsInWorld[2] = new Object(100, 50, 100, 55, true, false, "ground");
 		objectsInWorld[3] = new Object(100, 105, 20, 100, true, false, "pillar");
-		objectsInWorld[4] = new Object(200, objectsInWorld[1].getY()-100, 100, 100, false, false, "seaweed");
+		objectsInWorld[4] = new Object(200, (int)(objectsInWorld[1].getY()-100), 100, 100, false, false, "seaweed");
 	}
 }
